feat: improved UserModel
